+++ src/problems/problem_0034.f90
@@ -4,7 +4,7 @@ contains
 
 module subroutine euler0034(problem)
   type(problem_type), intent(inout) :: problem
-  integer(int32) :: i, sln
+  integer(int64) :: i, sln
 
   sln = 0
   do i = 1, 40585
@@ -14,9 +14,9 @@ module subroutine euler0034(problem)
 end subroutine euler0034
 
 elemental logical function is_curious(n)
-  integer(int32), intent(in) :: n
-  integer(int32) :: i, tmp
-  integer(int32), allocatable :: arr(:)
+  integer(int64), intent(in) :: n
+  integer(int64) :: i, tmp
+  integer(int64), allocatable :: arr(:)
 
   arr = to_array(n)
   tmp = 0
@@ -26,10 +26,10 @@ elemental logical function is_curious(n)
   is_curious = tmp == n
 end function is_curious
 
-elemental integer(int32) function factorial(n)
-  integer(int32), intent(in) :: n
+elemental integer(int64) function factorial(n)
+  integer(int64), intent(in) :: n
 
-  factorial = int(gamma(real(n + 1)), int32)
+  factorial = int(gamma(real(n + 1)), int64)
 end function factorial
 
 end submodule submodule_euler0034
