+++ src/problems/problem_0057.f90
@@ -6,7 +6,7 @@ module subroutine euler0057(problem)
   type(problem_type), intent(inout) :: problem
   real(real64), parameter :: delta = log10(1._real64 + sqrt(2._real64))
   real(real64) :: logp, logq
-  integer(int32) :: sln, i
+  integer(int64) :: sln, i
 
   logp = delta - log10(2._real64)
   logq = delta - log10(2._real64)*3.0/2.0
